+++ Changelog.md
@@ -9,3 +9,4 @@ Bugfixes:
 * Fix song search with special characters in title (#105)
 * In setlist-based song statistics, only count title once per setlist (#102)
 * Button color standardization in song view mode switch (#99)
+* Standardized and extended additional information that can be shown on slides: Song Books, Author, Publisher, Copyright and CCLI ID (see Program Settings dialog -> Master formatting). (#100, #101)